Split Backward trait and introduce composition
- Backward split
We have decided to remove the `no_grad()` and `with_grad()` methods
from the defition of the `Backward` trait, because those ones are not
strictly related to the backward propagation logic of the gradients,
but instead are required to implement some optimizations.
We now have the new `SwitchTensor` trait, containing the two methods
mentioned above, which can be used to define all the optimizations
desired.

- Composition
In order to remove a huge amount of duplicated code and simplify the
implementation of many nodes, we have decided to change the way the
`Backward` nodes are constructed. So far, for example, we had
`MatrixMatrixMulBackward` which contained all the code wrote for its
`Left` and `Right` versions, but from now on it will simply contain
these two nodes, delegating the calls to `.backward()` to them as well.
This applies to many other nodes, so the maintainability of the source
code has improved a lot, but careful testing is needed to ensure the
safety of this transition.