Reorganize modules and simplify Forward/Backward
- Reorganization of modules
All modules inside `node` have been placed on the same level, as it was
a bit tedious to write the same `use` many times in the various levels.
Obviously, this may not be the final organization.

- Forward
We have decided to remove the `was_computed` and `reset_computation`
functions that were contained in this trait, since the same logic can
be implemented on `Var`. In this way we managed to remove some
duplicated code and a bunch of dynamically dispatched function calls,
something that we hope will also improve the efficiency of the library.
Finally, we removed some unnecessary code in the `forward` method of
`Var`.

- Backward
We are no longer using two different `History` to store the `Backward`
and `Switch` pointers. Furthermore, we removed some unnecessary code in
the `backward` method of `VarDiff`.