Add `addition` tests and fix `convolution`
- Addition tests
The first version of the tests for the `addition` operation has been
implemented, but it may be reworked later. I decided to use the
`Result` approach, instead of the usual `assert`, because in this way
we can also propagate those returned from the `ndarray` library using
the `?` operator.

- Convolution fix
I removed some unnecessary `&` from the `convolution` module which were
highlighted by `Clippy`.

- Changed `Gradient` constructor
We can now create an object by providing a `ShapeBuilder` to its `new`
function, so that its construction is easier, especially when writing
tests.